Deep refactoring of the application and introduction of multiThreads.
Helper folder rename to internalUser:
Add the package-info.
Creation of the InternalUserFactory class, which gets whole part about creating the in memory pool of users from TourGuideService.

TourGuideService class:
move to UserService class of getUserRewards, getUserLocation, getUserByUsername, getAllUsers, addUser and trackUserLocation methods.
move to InternalUserFactory of whole part about creating the in memory pool of users.
move to UserService, the Tracker call.

UserService class:
Creation of the UserService class, which whole methods, getUserRewards, getUserLocation, getUserByUsername, getAllUsers, addUser and trackUserLocation methods from the TourGuideService class.
Inner class, TrackUserCallable who implement Callable to call in a multiThreads the GpsUtil lib.

RewardsService:
Creation of an inner class, CalculateUserRewardsCallable who implement Callable to perform in a multiThreads the UserRewards calculate.
Add a method calculateAllUsersRewards to calculate the UserRewards of a list of users.
Modify the calculateUserRewards to work with multiThreads.

Tracker:
Update the class cause of refactoring of TourGuideService and UserService.

TourGuideModule:
Update the class cause of refactoring of RewardsService.

DtoService:
Update the class cause of refactoring of TourGuideService and UserService.

User:
Modify addUserReward to save a simple UserReward or a List of UserRewards.
Add a verification to the addUserReward method.

JavaDoc:
Add or update javaDoc comments in all classes.

Test:
Update tests cause of the global refactoring of the application.